Fix bugs 31235, 31236
Tackles the edge cases of folders with GUID's as names and stop words as
names. For the former, we now silence fault exceptions made in
TryGetFolderById - which occur when we try to get folders of a
nonexistent GUID - and continue the workflow. For the latter, we now
check every folder by name if the search query-matching comparisons
fail.